Migrate from ant to gradle build system
There is a need of such a change because the apk does not work, but it is
nearly impossible to build a Android ant project now (not officially supported).

The font files are removed because I don't find them has any use in this sample
project.

There is an issue: payment does not work.  It only shows 'Press Here to pay'.  I
will take a look later.